Refactor SpecTests * Expose `LibYamlEventStream` on public surface. * Rename `SpecTests` to `ParserSpecTests`. * Add `SpectTestData` to accommodate emitter/serializer tests.
